chore: add Exception on when delete user

+++ src/main/java/com/project/paswordapi/Service/UserService.java
@@ -1,11 +1,15 @@
 package com.project.paswordapi.Service;
 
 import com.project.paswordapi.Entity.UserEntity;
+import com.project.paswordapi.Exception.DatabaseAccessException;
+import com.project.paswordapi.Exception.UserNotFoundException;
 import com.project.paswordapi.Repository.UserRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -17,7 +21,17 @@ public class UserService {
         userRepository.saveAll(userEntity);
     }
 
-    public void deleteUser(UUID userID){
-        userRepository.deleteById(userID);
+    public void deleteUser(UUID userID) throws UserNotFoundException, DatabaseAccessException {
+        try {
+            Optional<UserEntity> optionalUser = userRepository.findById(userID);
+            if (optionalUser.isPresent()) {
+                userRepository.deleteById(userID);
+            } else {
+                throw new UserNotFoundException("User with ID " + userID + " does not exist.");
+            }
+        } catch (IllegalArgumentException | DataAccessException e) {
+            throw new DatabaseAccessException("Error while deleting user with ID " + userID);
+        }
     }
+
 }
